extracts GetRecordResourceOperations.
